Fixes install
